Update package versions and enhance method documentation
Updated the version numbers of the `BlazorExpress.Bulma` and `BlazorExpress.Bulma.Docx` packages in the `BlazorExpress.ChartJS.Demo.RCL.csproj` file from `0.1.0` to `0.1.1` and from `0.2.0` to `0.2.1`, respectively. Also updated the version number of the `BlazorExpress.Core` package in the `BlazorExpress.ChartJS.csproj` file from `0.1.0` to `0.1.1`.

Added the `MethodReturnTypeName` attribute to several asynchronous methods in the `BarChart`, `DoughnutChart`, `LineChart`, `PieChart`, `PolarAreaChart`, `RadarChart`, and `ScatterChart` classes, enhancing the documentation by specifying the return type of the methods.

NOTE: This commit message is auto-generated using GitHub Copilot.